refactor: 파일이름 변경, jsDoc 추가
